Add JS and Solidity auto linter/prettier

+++ contracts/Accountant.sol
@@ -11,34 +11,48 @@ import "@openzeppelin/contracts-ethereum-package/contracts/math/SafeMath.sol";
 library Accountant {
   using SafeMath for uint256;
 
-  uint public constant interestDecimals = 1e18;
-  uint public constant blocksPerDay = 5760;
-  uint public constant blocksPerYear = (blocksPerDay * 365);
+  uint256 public constant interestDecimals = 1e18;
+  uint256 public constant blocksPerDay = 5760;
+  uint256 public constant blocksPerYear = (blocksPerDay * 365);
 
   struct PaymentAllocation {
-    uint interestPayment;
-    uint principalPayment;
-    uint additionalBalancePayment;
+    uint256 interestPayment;
+    uint256 principalPayment;
+    uint256 additionalBalancePayment;
   }
 
-  function calculateInterestAndPrincipalAccrued(CreditLine cl, uint blockNumber) public view returns(uint, uint) {
-    uint totalPayment = calculateAnnuityPayment(cl.balance(), cl.interestApr(), cl.termInDays(), cl.paymentPeriodInDays());
-    uint interestAccrued = calculateInterestAccrued(cl, blockNumber);
-    uint principalAccrued = calculatePrincipalAccrued(cl, totalPayment, interestAccrued, blockNumber);
+  function calculateInterestAndPrincipalAccrued(CreditLine cl, uint256 blockNumber)
+    public
+    view
+    returns (uint256, uint256)
+  {
+    uint256 totalPayment = calculateAnnuityPayment(
+      cl.balance(),
+      cl.interestApr(),
+      cl.termInDays(),
+      cl.paymentPeriodInDays()
+    );
+    uint256 interestAccrued = calculateInterestAccrued(cl, blockNumber);
+    uint256 principalAccrued = calculatePrincipalAccrued(cl, totalPayment, interestAccrued, blockNumber);
     return (interestAccrued, principalAccrued);
   }
 
-  function calculatePrincipalAccrued(CreditLine cl, uint periodPayment, uint interestAccrued, uint blockNumber) public view returns(uint) {
-    uint blocksPerPaymentPeriod = blocksPerDay * cl.paymentPeriodInDays();
+  function calculatePrincipalAccrued(
+    CreditLine cl,
+    uint256 periodPayment,
+    uint256 interestAccrued,
+    uint256 blockNumber
+  ) public view returns (uint256) {
+    uint256 blocksPerPaymentPeriod = blocksPerDay * cl.paymentPeriodInDays();
     // Math.min guards against overflow. See comment in the calculateInterestAccrued for further explanation.
-    uint lastUpdatedBlock = Math.min(blockNumber, cl.lastUpdatedBlock());
-    uint numBlocksElapsed = blockNumber.sub(lastUpdatedBlock);
+    uint256 lastUpdatedBlock = Math.min(blockNumber, cl.lastUpdatedBlock());
+    uint256 numBlocksElapsed = blockNumber.sub(lastUpdatedBlock);
     int128 fractionOfPeriod = FPMath.divi(int256(numBlocksElapsed), int256(blocksPerPaymentPeriod));
-    uint periodPaymentFraction = uint(FPMath.muli(fractionOfPeriod, int256(periodPayment)));
+    uint256 periodPaymentFraction = uint256(FPMath.muli(fractionOfPeriod, int256(periodPayment)));
     return periodPaymentFraction.sub(interestAccrued);
   }
 
-  function calculateInterestAccrued(CreditLine cl, uint blockNumber) public view returns(uint) {
+  function calculateInterestAccrued(CreditLine cl, uint256 blockNumber) public view returns (uint256) {
     // We use Math.min here to prevent integer overflow (ie. go negative) when calculating
     // numBlocksElapsed. Typically this shouldn't be possible, because
     // the lastUpdatedBlock couldn't be *after* the current blockNumber. However, when assessing
@@ -47,14 +61,19 @@ library Accountant {
     // This use of min should not generate incorrect interest calculations, since
     // this functions purpose is just to normalize balances, and  will be called any time
     // a balance affecting action takes place (eg. drawdown, repayment, assessment)
-    uint lastUpdatedBlock = Math.min(blockNumber, cl.lastUpdatedBlock());
+    uint256 lastUpdatedBlock = Math.min(blockNumber, cl.lastUpdatedBlock());
 
-    uint numBlocksElapsed = blockNumber.sub(lastUpdatedBlock);
-    uint totalInterestPerYear = (cl.balance().mul(cl.interestApr())).div(interestDecimals);
+    uint256 numBlocksElapsed = blockNumber.sub(lastUpdatedBlock);
+    uint256 totalInterestPerYear = (cl.balance().mul(cl.interestApr())).div(interestDecimals);
     return totalInterestPerYear.mul(numBlocksElapsed).div(blocksPerYear);
   }
 
-  function calculateAnnuityPayment(uint balance, uint interestApr, uint termInDays, uint paymentPeriodInDays) public pure returns(uint) {
+  function calculateAnnuityPayment(
+    uint256 balance,
+    uint256 interestApr,
+    uint256 termInDays,
+    uint256 paymentPeriodInDays
+  ) public pure returns (uint256) {
     /*
     This is the standard amortization formula for an annuity payment amount.
     See: https://en.wikipedia.org/wiki/Amortization_calculator
@@ -67,9 +86,8 @@ library Accountant {
     doing the formula above, and then turning it back into an int64 at the end.
     */
 
-
     // Components used in the formula
-    uint periodsPerTerm = termInDays / paymentPeriodInDays;
+    uint256 periodsPerTerm = termInDays / paymentPeriodInDays;
     int128 one = FPMath.fromInt(int256(1));
     int128 annualRate = FPMath.divi(int256(interestApr), int256(interestDecimals));
     int128 dailyRate = FPMath.div(annualRate, FPMath.fromInt(int256(365)));
@@ -81,27 +99,32 @@ library Accountant {
       return balance / periodsPerTerm;
     }
     int128 paymentFractionFP = FPMath.div(periodRate, denominator);
-    uint paymentFraction = uint(FPMath.muli(paymentFractionFP, int256(1e18)));
+    uint256 paymentFraction = uint256(FPMath.muli(paymentFractionFP, int256(1e18)));
 
     return (balance * paymentFraction) / 1e18;
   }
 
-  function allocatePayment(uint paymentAmount, uint balance, uint interestOwed, uint principalOwed) public pure returns(PaymentAllocation memory) {
-    uint paymentRemaining = paymentAmount;
-    uint interestPayment = Math.min(interestOwed, paymentRemaining);
+  function allocatePayment(
+    uint256 paymentAmount,
+    uint256 balance,
+    uint256 interestOwed,
+    uint256 principalOwed
+  ) public pure returns (PaymentAllocation memory) {
+    uint256 paymentRemaining = paymentAmount;
+    uint256 interestPayment = Math.min(interestOwed, paymentRemaining);
     paymentRemaining = paymentRemaining.sub(interestPayment);
 
-    uint principalPayment = Math.min(principalOwed, paymentRemaining);
+    uint256 principalPayment = Math.min(principalOwed, paymentRemaining);
     paymentRemaining = paymentRemaining.sub(principalPayment);
 
-    uint balanceRemaining = balance.sub(principalPayment);
-    uint additionalBalancePayment = Math.min(paymentRemaining, balanceRemaining);
+    uint256 balanceRemaining = balance.sub(principalPayment);
+    uint256 additionalBalancePayment = Math.min(paymentRemaining, balanceRemaining);
 
-    return PaymentAllocation({
-      interestPayment: interestPayment,
-      principalPayment: principalPayment,
-      additionalBalancePayment: additionalBalancePayment
-    });
+    return
+      PaymentAllocation({
+        interestPayment: interestPayment,
+        principalPayment: principalPayment,
+        additionalBalancePayment: additionalBalancePayment
+      });
   }
-
 }
