+++ contracts/CreditDesk.sol
@@ -15,13 +15,13 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
   using SafeMath for uint256;
 
   // Approximate number of blocks
-  uint public constant blocksPerDay = 5760;
+  uint256 public constant blocksPerDay = 5760;
   address public poolAddress;
-  uint public maxUnderwriterLimit = 0;
-  uint public transactionLimit = 0;
+  uint256 public maxUnderwriterLimit = 0;
+  uint256 public transactionLimit = 0;
 
   struct Underwriter {
-    uint governanceLimit;
+    uint256 governanceLimit;
     address[] creditLines;
   }
 
@@ -29,13 +29,19 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     address[] creditLines;
   }
 
-  event PaymentMade(address indexed payer, address indexed creditLine, uint interestAmount, uint principalAmount, uint remainingAmount);
-  event PrepaymentMade(address indexed payer, address indexed creditLine, uint prepaymentAmount);
-  event DrawdownMade(address indexed borrower, address indexed creditLine, uint drawdownAmount);
+  event PaymentMade(
+    address indexed payer,
+    address indexed creditLine,
+    uint256 interestAmount,
+    uint256 principalAmount,
+    uint256 remainingAmount
+  );
+  event PrepaymentMade(address indexed payer, address indexed creditLine, uint256 prepaymentAmount);
+  event DrawdownMade(address indexed borrower, address indexed creditLine, uint256 drawdownAmount);
   event CreditLineCreated(address indexed borrower, address indexed creditLine);
   event PoolAddressUpdated(address indexed oldAddress, address indexed newAddress);
-  event GovernanceUpdatedUnderwriterLimit(address indexed underwriter, uint newLimit);
-  event LimitChanged(address indexed owner, string limitType, uint amount);
+  event GovernanceUpdatedUnderwriterLimit(address indexed underwriter, uint256 newLimit);
+  event LimitChanged(address indexed owner, string limitType, uint256 amount);
 
   mapping(address => Underwriter) public underwriters;
   mapping(address => Borrower) private borrowers;
@@ -46,7 +52,7 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     setPoolAddress(_poolAddress);
   }
 
-  function setUnderwriterGovernanceLimit(address underwriterAddress, uint limit) external onlyOwner whenNotPaused {
+  function setUnderwriterGovernanceLimit(address underwriterAddress, uint256 limit) external onlyOwner whenNotPaused {
     Underwriter storage underwriter = underwriters[underwriterAddress];
     require(withinMaxUnderwriterLimit(limit), "This limit is greater than the max allowed by the protocol");
     underwriter.governanceLimit = limit;
@@ -55,26 +61,34 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
 
   function createCreditLine(
     address _borrower,
-    uint _limit,
-    uint _interestApr,
-    uint _minCollateralPercent,
-    uint _paymentPeriodInDays,
-    uint _termInDays
+    uint256 _limit,
+    uint256 _interestApr,
+    uint256 _minCollateralPercent,
+    uint256 _paymentPeriodInDays,
+    uint256 _termInDays
   ) external whenNotPaused {
     Underwriter storage underwriter = underwriters[msg.sender];
     Borrower storage borrower = borrowers[_borrower];
     require(underwriterCanCreateThisCreditLine(_limit, underwriter), "The underwriter cannot create this credit line");
 
     CreditLine cl = new CreditLine();
-    cl.initialize(_borrower, msg.sender, _limit, _interestApr, _minCollateralPercent, _paymentPeriodInDays, _termInDays);
+    cl.initialize(
+      _borrower,
+      msg.sender,
+      _limit,
+      _interestApr,
+      _minCollateralPercent,
+      _paymentPeriodInDays,
+      _termInDays
+    );
     cl.authorizePool(poolAddress);
 
     underwriter.creditLines.push(address(cl));
     borrower.creditLines.push(address(cl));
     emit CreditLineCreated(_borrower, address(cl));
-	}
+  }
 
-  function drawdown(uint amount, address creditLineAddress) external whenNotPaused {
+  function drawdown(uint256 amount, address creditLineAddress) external whenNotPaused {
     CreditLine cl = CreditLine(creditLineAddress);
     require(cl.borrower() == msg.sender, "You do not belong to this credit line");
     // Not strictly necessary, but provides a better error message to the user
@@ -86,8 +100,8 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
       cl.setTermEndBlock(calculateNewTermEndBlock(cl));
       cl.setNextDueBlock(calculateNextDueBlock(cl));
     }
-    (uint interestOwed, uint principalOwed) = getInterestAndPrincipalOwedAsOf(cl, block.number);
-    uint balance = cl.balance().add(amount);
+    (uint256 interestOwed, uint256 principalOwed) = getInterestAndPrincipalOwedAsOf(cl, block.number);
+    uint256 balance = cl.balance().add(amount);
 
     updateCreditLineAccounting(cl, balance, interestOwed, principalOwed);
     getPool().transferFrom(poolAddress, msg.sender, amount);
@@ -95,14 +109,19 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     emit DrawdownMade(msg.sender, address(cl), amount);
   }
 
-  function pay(address creditLineAddress, uint amount) external payable whenNotPaused {
+  function pay(address creditLineAddress, uint256 amount) external payable whenNotPaused {
     CreditLine cl = CreditLine(creditLineAddress);
 
     require(withinTransactionLimit(amount), "Amount is over the per-transaction limit");
     // Not strictly necessary, but provides a faster/better error message to the user
     require(getPool().enoughBalance(msg.sender, amount), "You have insufficent balance for this payment");
 
-    (uint paymentRemaining, uint interestPayment, uint principalPayment) = handlePayment(cl, amount, block.number, true);
+    (uint256 paymentRemaining, uint256 interestPayment, uint256 principalPayment) = handlePayment(
+      cl,
+      amount,
+      block.number,
+      true
+    );
     if (paymentRemaining > 0) {
       getPool().transferFrom(msg.sender, creditLineAddress, paymentRemaining);
       cl.setCollateralBalance(cl.collateralBalance().add(paymentRemaining));
@@ -117,23 +136,23 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     emit PaymentMade(cl.borrower(), address(cl), interestPayment, principalPayment, paymentRemaining);
   }
 
-  function prepay(address payable creditLineAddress, uint amount) external payable whenNotPaused {
+  function prepay(address payable creditLineAddress, uint256 amount) external payable whenNotPaused {
     CreditLine cl = CreditLine(creditLineAddress);
 
     require(withinTransactionLimit(amount), "Amount is over the per-transaction limit");
 
     getPool().transferFrom(msg.sender, creditLineAddress, amount);
-    uint newPrepaymentBalance = cl.prepaymentBalance().add(amount);
+    uint256 newPrepaymentBalance = cl.prepaymentBalance().add(amount);
     cl.setPrepaymentBalance(newPrepaymentBalance);
 
     emit PrepaymentMade(msg.sender, address(cl), amount);
   }
 
-  function addCollateral(address payable creditLineAddress, uint amount) external payable whenNotPaused {
+  function addCollateral(address payable creditLineAddress, uint256 amount) external payable whenNotPaused {
     CreditLine cl = CreditLine(creditLineAddress);
 
     getPool().transferFrom(msg.sender, creditLineAddress, amount);
-    uint newCollateralBalance = cl.collateralBalance().add(amount);
+    uint256 newCollateralBalance = cl.collateralBalance().add(amount);
     cl.setCollateralBalance(newCollateralBalance);
   }
 
@@ -144,7 +163,12 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
       return;
     }
 
-    (uint paymentRemaining, uint interestPayment, uint principalPayment) = handlePayment(cl, cl.prepaymentBalance(), cl.nextDueBlock(), false);
+    (uint256 paymentRemaining, uint256 interestPayment, uint256 principalPayment) = handlePayment(
+      cl,
+      cl.prepaymentBalance(),
+      cl.nextDueBlock(),
+      false
+    );
 
     cl.setPrepaymentBalance(paymentRemaining);
     getPool().collectInterestRepayment(msg.sender, interestPayment);
@@ -164,17 +188,17 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     return poolAddress = newPoolAddress;
   }
 
-  function setMaxUnderwriterLimit(uint amount) public onlyOwner whenNotPaused {
+  function setMaxUnderwriterLimit(uint256 amount) public onlyOwner whenNotPaused {
     maxUnderwriterLimit = amount;
     emit LimitChanged(msg.sender, "maxUnderwriterLimit", amount);
   }
 
-  function setTransactionLimit(uint amount) public onlyOwner whenNotPaused {
+  function setTransactionLimit(uint256 amount) public onlyOwner whenNotPaused {
     transactionLimit = amount;
     emit LimitChanged(msg.sender, "transactionLimit", amount);
   }
 
-  function setPoolTotalFundsLimit(uint amount) public onlyOwner whenNotPaused {
+  function setPoolTotalFundsLimit(uint256 amount) public onlyOwner whenNotPaused {
     getPool().setTotalFundsLimit(amount);
   }
 
@@ -190,20 +214,42 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
 
   /*
    * Internal Functions
-  */
+   */
 
-  function handlePayment(CreditLine cl, uint paymentAmount, uint asOfBlock, bool allowFullBalancePayOff) internal returns (uint, uint, uint) {
-    (uint interestOwed, uint principalOwed) = getInterestAndPrincipalOwedAsOf(cl, asOfBlock);
-    Accountant.PaymentAllocation memory pa = Accountant.allocatePayment(paymentAmount, cl.balance(), interestOwed, principalOwed);
+  function handlePayment(
+    CreditLine cl,
+    uint256 paymentAmount,
+    uint256 asOfBlock,
+    bool allowFullBalancePayOff
+  )
+    internal
+    returns (
+      uint256,
+      uint256,
+      uint256
+    )
+  {
+    (uint256 interestOwed, uint256 principalOwed) = getInterestAndPrincipalOwedAsOf(cl, asOfBlock);
+    Accountant.PaymentAllocation memory pa = Accountant.allocatePayment(
+      paymentAmount,
+      cl.balance(),
+      interestOwed,
+      principalOwed
+    );
 
-    uint newBalance = cl.balance().sub(pa.principalPayment);
+    uint256 newBalance = cl.balance().sub(pa.principalPayment);
     if (allowFullBalancePayOff) {
       newBalance = newBalance.sub(pa.additionalBalancePayment);
     }
-    uint totalPrincipalPayment = cl.balance().sub(newBalance);
-    uint paymentRemaining = paymentAmount.sub(pa.interestPayment).sub(totalPrincipalPayment);
+    uint256 totalPrincipalPayment = cl.balance().sub(newBalance);
+    uint256 paymentRemaining = paymentAmount.sub(pa.interestPayment).sub(totalPrincipalPayment);
 
-    updateCreditLineAccounting(cl, newBalance, interestOwed.sub(pa.interestPayment), principalOwed.sub(pa.principalPayment));
+    updateCreditLineAccounting(
+      cl,
+      newBalance,
+      interestOwed.sub(pa.interestPayment),
+      principalOwed.sub(pa.principalPayment)
+    );
 
     assert(paymentRemaining.add(pa.interestPayment).add(totalPrincipalPayment) == paymentAmount);
 
@@ -218,26 +264,33 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     return Pool(poolAddress);
   }
 
-  function getInterestAndPrincipalOwedAsOf(CreditLine cl, uint blockNumber) internal view returns (uint, uint) {
-    (uint interestAccrued, uint principalAccrued) = Accountant.calculateInterestAndPrincipalAccrued(cl, blockNumber);
+  function getInterestAndPrincipalOwedAsOf(CreditLine cl, uint256 blockNumber)
+    internal
+    view
+    returns (uint256, uint256)
+  {
+    (uint256 interestAccrued, uint256 principalAccrued) = Accountant.calculateInterestAndPrincipalAccrued(
+      cl,
+      blockNumber
+    );
     return (cl.interestOwed().add(interestAccrued), cl.principalOwed().add(principalAccrued));
   }
 
-  function withinCreditLimit(uint amount, CreditLine cl) internal view returns(bool) {
+  function withinCreditLimit(uint256 amount, CreditLine cl) internal view returns (bool) {
     return cl.balance().add(amount) <= cl.limit();
   }
 
-  function withinTransactionLimit(uint amount) internal view returns(bool) {
+  function withinTransactionLimit(uint256 amount) internal view returns (bool) {
     return amount <= transactionLimit;
   }
 
-  function calculateNewTermEndBlock(CreditLine cl) internal view returns (uint) {
+  function calculateNewTermEndBlock(CreditLine cl) internal view returns (uint256) {
     return block.number.add(blocksPerDay.mul(cl.termInDays()));
   }
 
-  function calculateNextDueBlock(CreditLine cl) internal view returns (uint) {
-    uint blocksPerPeriod = cl.paymentPeriodInDays().mul(blocksPerDay);
-    uint currentNextDueBlock;
+  function calculateNextDueBlock(CreditLine cl) internal view returns (uint256) {
+    uint256 blocksPerPeriod = cl.paymentPeriodInDays().mul(blocksPerDay);
+    uint256 currentNextDueBlock;
     if (cl.nextDueBlock() != 0) {
       currentNextDueBlock = cl.nextDueBlock();
     } else {
@@ -246,26 +299,35 @@ contract CreditDesk is Initializable, OwnableUpgradeSafe, OwnerPausable {
     return currentNextDueBlock.add(blocksPerPeriod);
   }
 
-  function underwriterCanCreateThisCreditLine(uint newAmount, Underwriter storage underwriter) internal view returns(bool) {
-    uint creditCurrentlyExtended = getCreditCurrentlyExtended(underwriter);
-    uint totalToBeExtended = creditCurrentlyExtended.add(newAmount);
+  function underwriterCanCreateThisCreditLine(uint256 newAmount, Underwriter storage underwriter)
+    internal
+    view
+    returns (bool)
+  {
+    uint256 creditCurrentlyExtended = getCreditCurrentlyExtended(underwriter);
+    uint256 totalToBeExtended = creditCurrentlyExtended.add(newAmount);
     return totalToBeExtended <= underwriter.governanceLimit;
   }
 
-  function withinMaxUnderwriterLimit(uint amount) internal view returns (bool) {
+  function withinMaxUnderwriterLimit(uint256 amount) internal view returns (bool) {
     return amount <= maxUnderwriterLimit;
   }
 
-  function getCreditCurrentlyExtended(Underwriter storage underwriter) internal view returns (uint) {
-    uint creditExtended = 0;
-    for (uint i = 0; i < underwriter.creditLines.length; i++) {
+  function getCreditCurrentlyExtended(Underwriter storage underwriter) internal view returns (uint256) {
+    uint256 creditExtended = 0;
+    for (uint256 i = 0; i < underwriter.creditLines.length; i++) {
       CreditLine cl = CreditLine(underwriter.creditLines[i]);
       creditExtended = creditExtended.add(cl.limit());
     }
     return creditExtended;
   }
 
-  function updateCreditLineAccounting(CreditLine cl, uint balance, uint interestOwed, uint principalOwed) internal {
+  function updateCreditLineAccounting(
+    CreditLine cl,
+    uint256 balance,
+    uint256 interestOwed,
+    uint256 principalOwed
+  ) internal {
     cl.setBalance(balance);
     cl.setInterestOwed(interestOwed);
     cl.setPrincipalOwed(principalOwed);
