+++ contracts/CreditLine.sol
@@ -11,31 +11,31 @@ contract CreditLine is Initializable, OwnableUpgradeSafe {
   // Credit line terms
   address public borrower;
   address public underwriter;
-  uint public collateral;
-  uint public limit;
-  uint public interestApr;
-  uint public minCollateralPercent;
-  uint public paymentPeriodInDays;
-  uint public termInDays;
+  uint256 public collateral;
+  uint256 public limit;
+  uint256 public interestApr;
+  uint256 public minCollateralPercent;
+  uint256 public paymentPeriodInDays;
+  uint256 public termInDays;
 
   // Accounting variables
-  uint public balance;
-  uint public interestOwed;
-  uint public principalOwed;
-  uint public prepaymentBalance;
-  uint public collateralBalance;
-  uint public termEndBlock;
-  uint public nextDueBlock;
-  uint public lastUpdatedBlock;
+  uint256 public balance;
+  uint256 public interestOwed;
+  uint256 public principalOwed;
+  uint256 public prepaymentBalance;
+  uint256 public collateralBalance;
+  uint256 public termEndBlock;
+  uint256 public nextDueBlock;
+  uint256 public lastUpdatedBlock;
 
   function initialize(
     address _borrower,
     address _underwriter,
-    uint _limit,
-    uint _interestApr,
-    uint _minCollateralPercent,
-    uint _paymentPeriodInDays,
-    uint _termInDays
+    uint256 _limit,
+    uint256 _interestApr,
+    uint256 _minCollateralPercent,
+    uint256 _paymentPeriodInDays,
+    uint256 _termInDays
   ) public initializer {
     __Ownable_init();
     borrower = _borrower;
@@ -48,39 +48,39 @@ contract CreditLine is Initializable, OwnableUpgradeSafe {
     lastUpdatedBlock = block.number;
   }
 
-  function setTermEndBlock(uint newTermEndBlock) external onlyOwner returns (uint) {
+  function setTermEndBlock(uint256 newTermEndBlock) external onlyOwner returns (uint256) {
     return termEndBlock = newTermEndBlock;
   }
 
-  function setNextDueBlock(uint newNextDueBlock) external onlyOwner returns (uint) {
+  function setNextDueBlock(uint256 newNextDueBlock) external onlyOwner returns (uint256) {
     return nextDueBlock = newNextDueBlock;
   }
 
-  function setBalance(uint newBalance) external onlyOwner returns(uint) {
+  function setBalance(uint256 newBalance) external onlyOwner returns (uint256) {
     return balance = newBalance;
   }
 
-  function setInterestOwed(uint newInterestOwed) external onlyOwner returns (uint) {
+  function setInterestOwed(uint256 newInterestOwed) external onlyOwner returns (uint256) {
     return interestOwed = newInterestOwed;
   }
 
-  function setPrincipalOwed(uint newPrincipalOwed) external onlyOwner returns (uint) {
+  function setPrincipalOwed(uint256 newPrincipalOwed) external onlyOwner returns (uint256) {
     return principalOwed = newPrincipalOwed;
   }
 
-  function setPrepaymentBalance(uint newPrepaymentBalance) external onlyOwner returns (uint) {
+  function setPrepaymentBalance(uint256 newPrepaymentBalance) external onlyOwner returns (uint256) {
     return prepaymentBalance = newPrepaymentBalance;
   }
 
-  function setCollateralBalance(uint newCollateralBalance) external onlyOwner returns (uint) {
+  function setCollateralBalance(uint256 newCollateralBalance) external onlyOwner returns (uint256) {
     return collateralBalance = newCollateralBalance;
   }
 
-  function setLastUpdatedBlock(uint newLastUpdatedBlock) external onlyOwner returns (uint) {
+  function setLastUpdatedBlock(uint256 newLastUpdatedBlock) external onlyOwner returns (uint256) {
     return lastUpdatedBlock = newLastUpdatedBlock;
   }
 
-  function setLimit(uint newAmount) external onlyOwnerOrUnderwriter returns (uint) {
+  function setLimit(uint256 newAmount) external onlyOwnerOrUnderwriter returns (uint256) {
     return limit = newAmount;
   }
 
@@ -88,7 +88,7 @@ contract CreditLine is Initializable, OwnableUpgradeSafe {
     address erc20address = Pool(poolAddress).erc20address();
 
     // Approve the pool for an infinite amount
-    ERC20UpgradeSafe(erc20address).approve(poolAddress, uint(-1));
+    ERC20UpgradeSafe(erc20address).approve(poolAddress, uint256(-1));
   }
 
   modifier onlyOwnerOrUnderwriter() {
